adding an event listener to the swichBtn to show or hide the color container
styling the input slider by using -webkit-appearance: none;
By giving it the value of none this tells each respective browser to clear out any default styles.
This allows us to be able to start from scratch and build up the look of the input from that point.

use  ::-webkit-slider-thumb
Targets the thumb in Chrome, Safari, and Edge Chromium.

The first style we need to apply to the
::-webkit-slider-thumb pseudo-element is the -webkit-appearance: none; vendor prefix.
We used this property within the “Baseline Styles” section to override the general
default styles that are applied by the browser and it serves a similar purpose on the thumb.

In order to properly center the thumb on the track we can use the following formula
 and apply it to the margin-top property:
    margin-top = (track height in pixels / 2) - (thumb height in pixels /2)
    margin-top = -7.5px

use ::-webkit-slider-runnable-track webkit will be applied to the Chrome, Safari, Opera and Edge
(post-Chromium) browsers. Anything prefixed with -moz pertains to Firefox.